+++ library/Validator.php
@@ -72,6 +72,7 @@ use Respect\Validation\Rules\Key;
  * @method static Validator key(string $reference, Validatable $referenceValidator = null, bool $mandatory = true)
  * @method static Validator keyNested(string $reference, Validatable $referenceValidator = null, bool $mandatory = true)
  * @method static Validator keySet(Key $rule...)
+ * @method static Validator keyValue(string $comparedKey, string $ruleName, string $baseKey)
  * @method static Validator leapDate(string $format)
  * @method static Validator leapYear()
  * @method static Validator length(int $min = null, int $max = null, bool $inclusive = true)