+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes of the Respect\Validation releases are documented in this fi
 - Create "Infinite" rule (#397)
 - Create "KeyNested" rule (#429)
 - Create "KeySet" rule (#374)
+- Create "KeyValue" rule (#441)
 - Create "Mimetype" rule (#361)
 - Create "Optional" rule (#423)
 - Create "ResourceType" rule (#397)